+++ packages/react-universal/CHANGELOG.md
@@ -1,3 +1,7 @@
+## Next
+
+* Export a `curiRoot()` (higher-order component) function to create a routing provider component instead of a `<CuriProvider>`. There should only be one Curi `router` in an application, so there shouldn't be a reason to swap routers.
+
 ## 1.0.0-beta.0
 
 * Migrated from `@curi/react` to `@curi/react-universal`.